+++ src/zem/plugins/coreutils/false.py
@@ -6,5 +6,5 @@ class FalseCommand(BaseCommand):
     help = "Do nothing, unsuccessfully"
     usage = "false"
 
-    def execute(self, args, context, stdin=None, stdout=None) -> int:
+    def execute(self, args, context, stdin=None, stdout=None, stderr=None) -> int:
         return 1